+++ main.go
@@ -14,7 +14,7 @@ import (
 )
 
 const (
-	Version = "1.8.3"
+	Version = "1.8.4"
 	Name    = "qmax-code"
 )